refactor(concurrency): delete parallel concurrent/ dir, single concurrency/ module
Two top-level dirs with overlapping names ("concurrent" + "concurrency")
was an architecture smell: neither was the canonical answer to "where
do concurrency primitives live?" Joel called the multi-source-of-truth
problem at the file-system level. This consolidates into one.

Per Joel directive 2026-05-18: "we have ZERO users we are in FULLBLOWN
rust driven dev" — directory duplications get the loser deleted in one
commit, no transition period.

## What this changes

Before:
  continuum-core/src/concurrent/
    mod.rs (11 LOC — re-exporter)
    message_processor.rs (129 LOC — MessageProcessor trait)
    priority_queue.rs (191 LOC — PriorityQueue trait)
  continuum-core/src/concurrency/
    mod.rs (629 LOC — ConcurrencyPolicy + TokioConcurrencyPolicy +
            single-flight maps + refcount guards)

After:
  continuum-core/src/concurrency/
    mod.rs (33 LOC — thin re-exporter, public API flat at
            crate::concurrency::*)
    policy.rs (629 LOC — moved from old concurrency/mod.rs)
    message_processor.rs (moved from concurrent/)
    priority_queue.rs (moved from concurrent/)

## Why concurrency/ wins

- Has the actual policy machinery used by real callers
  (cognition::shared_analysis + live::transport::livekit_agent)
- The name describes the broader concern (policies + data structures)
- concurrent/ had no current callers outside its own re-exports

## Caller updates

- src/lib.rs: drop `pub mod concurrent;` + `pub use concurrent::*;`
  (replaced with `pub use concurrency::*;` so flat re-exports still
  resolve crate::MessageProcessor / crate::PriorityQueue)
- src/logging/mod.rs: log-routing prefix match `concurrent::` →
  `concurrency::`; log dir `system/concurrent` → `system/concurrency`

## Verification

- cargo check --features metal,accelerate: clean (50 pre-existing warnings,
  none in this change)
- cargo test cognition::shared_analysis::: green (single-flight code path
  unchanged)
- cargo test concurrency::: 7 passed, 0 failed, 1 ignored
- No new test scaffolding needed; this is pure file moves + import edits.

Ref: claude-tab-1 broad-code-audit broadcast 2026-05-18 19:40Z naming this
as one of the parallel-dir smells; Joel directive 2026-05-18 19:44Z on
zero-users full-blown-Rust-dev mode (no migration ceremony).
