+++ src/workers/continuum-core/src/concurrency/message_processor.rs
