--- src/workers/continuum-core/src/concurrent/mod.rs
@@ -1,11 +1,0 @@
-//! Reusable concurrent patterns for message processing
-//!
-//! OOP-style traits for common operations:
-//! - PriorityQueue<T>: Generic priority-based message queue
-//! - MessageProcessor<T>: Process messages concurrently
-//! - EventBus<T>: Publish-subscribe pattern
-pub mod message_processor;
-pub mod priority_queue;
-
-pub use message_processor::*;
-pub use priority_queue::*;
